requirements.txt
pip~=25.0.1
attrs~=25.1.0
PySocks~=1.7.1
pycryptodomex~=3.21.0
SQLAlchemy~=2.0.38